% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_annotation.R
\docType{methods}
\name{annotateInteractions}
\alias{annotateInteractions}
\alias{annotateInteractions,GenomicInteractions,list-method}
\title{Annotate interactions}
\usage{
annotateInteractions(GIObject, annotations)

\S4method{annotateInteractions}{GenomicInteractions,list}(GIObject,
  annotations)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object to be annotated.}

\item{annotations}{A list containing \linkS4class{GRanges} (or \linkS4class{GRangesList}) objects to be used to annotate \code{GIObject}.}
}
\value{
\code{NULL} is invisibly returned, 
while the value passed as \code{GIObject} is modified in the global namespace.
}
\description{
This function will annotate the interactions in a \linkS4class{GenomicInteractions} object with node class information.
}
\details{
For each interaction, this function identifies the entries of \code{annotations} that overlap an anchor region.
For each entry in \code{annotations}, a metadata column is added to the \linkS4class{GenomicInteractions} object,
containing a list that specifies the elements inside the entry that overlap each interaction.
The metadata column is named as \code{NAME.id} where \code{NAME} is the name of the entry.
Anonymous lists will be given names \code{FEATURE#.id} where \code{#} is the position in the list.

For each anchor, a \code{node.class} metadata column will also be added, 
containing the name of the list element which was \emph{first} annotated to each range.
Ranges with no overlaps will be classified as \code{"distal"}. 
The identifiers for each individual feature/annotation are taken from either the name of the list item in the case of a GRangesList or from either the names of a the provided GRanges or an id column in its associated metadata.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)
mcols(regions(hic_example_data, type=1))

data(mm9_refseq_promoters)
mm9_refseq_grl <- split(mm9_refseq_promoters, mm9_refseq_promoters$id)
annotateInteractions(hic_example_data, list(promoter=mm9_refseq_grl))
mcols(regions(hic_example_data, type=1))

}
