% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_annotation.R
\docType{methods}
\name{annotateRegions}
\alias{annotateRegions}
\alias{annotateRegions,GenomicInteractions,character,vector-method}
\title{Annotate regions}
\usage{
annotateRegions(GIObject, name, dat)

\S4method{annotateRegions}{GenomicInteractions,character,vector}(GIObject,
  name, dat)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object.}

\item{name}{Character vector to be used as the column name.}

\item{dat}{Vector of the same length as \code{regions(GIObject, 1)},
containing data with which to annotate the object.}
}
\value{
\code{NULL} is invisibly returned, 
while the value passed as \code{GIObject} is modified in the global namespace.
}
\description{
This function adds metadata parallel to the \code{regions} slot of a \linkS4class{GenomicInteractions} object.
}
\details{
This function only works if both \code{regions(GIObject, 1)} and \code{regions(GIObject, 2)} are identical.
In cases where they are not identical, it doesn't make sense to assign a single \code{dat} vector to the metadata.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)
mcols(regions(hic_example_data, type=1))

chip <- runif(length(regions(hic_example_data, type=1)), max=1000)
annotateRegions(hic_example_data, "chip", chip)
mcols(regions(hic_example_data, type=1))

}
