% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\docType{methods}
\name{asBED,GenomicInteractions-method}
\alias{asBED,GenomicInteractions-method}
\alias{asBED}
\title{Coerce to BED structure}
\usage{
\S4method{asBED}{GenomicInteractions}(x, keep.mcols = FALSE,
  score = "score")
}
\arguments{
\item{x}{A \linkS4class{GenomicInteractions} object.}

\item{keep.mcols}{Logical scalar, specifying whether to keep non-BED12 columns in the output object.}

\item{score}{String specifying which field to export as the \code{score} field in BED12.
Defaults to "auto" which will choose \code{"score"}, then \code{"counts"}, if present, or fill the column with zeros.}
}
\value{
A \linkS4class{GRanges} object with the metadata columns \code{name}, \code{blockStarts} and \code{blockSizes}.
}
\description{
Coerce the structure of a \linkS4class{GenomicInteractions} into a \linkS4class{GRanges} following BED-like conventions, 
i.e., with columns for blocks and thick regions.
}
\details{
Note that setting \code{keep.mcols=TRUE} may cause problems with some parsers.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)
asBED(hic_example_data)
}
\author{
Malcolm Perry, Elizabeth Ing-Simmons
}
