% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionTrack-class.R
\name{availableDisplayPars}
\alias{availableDisplayPars}
\title{Available display parameters}
\usage{
availableDisplayPars(class)
}
\arguments{
\item{class}{A valid track object class name, or the object itself, in which case the class name is determined from it.}
}
\value{
A named list of the default display parameters.
}
\description{
Query the default display parameters for an \linkS4class{InteractionTrack} object.
}
\details{
This function provides the same functionality as \code{Gviz::availableDisplayPars} and allows the user to display the default display parameters for the \code{InteractionTrack} class. 
If the class of the track is not \code{"InteractionTrack"} then the function calls the \code{\link{availableDisplayPars}} method in \pkg{Gviz}.
}
\examples{
availableDisplayPars("InteractionTrack")

}
