% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummaryStats.R
\name{categoriseInteractions}
\alias{categoriseInteractions}
\title{Categorize interaction types}
\usage{
categoriseInteractions(GIObject, node.classes = NULL,
  viewpoints = NULL)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object.}

\item{node.classes}{Character vector specifying the node classes to include in the plot.
Defaults to all node classes.}

\item{viewpoints}{Character vector of acceptable node classes,
where interactions are only considered if least one anchor is of an acceptable node class. 
Defaults to all classes in \code{node.classes}.}
}
\value{
A data.frame.
}
\description{
Get the numbers of interaction types existing in a \linkS4class{GenomicInteractions} object.
}
\examples{
library("GenomicRanges")
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)

data(mm9_refseq_promoters)
mm9_refseq_grl = split(mm9_refseq_promoters, mm9_refseq_promoters$id)
annotateInteractions(hic_example_data, list(promoter=mm9_refseq_grl))
categoriseInteractions(hic_example_data)
}
