% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/processing.R
\docType{methods}
\name{countsBetweenAnchors}
\alias{countsBetweenAnchors}
\alias{countsBetweenAnchors,GenomicInteractions,GRanges-method}
\title{Summarise interactions between anchors}
\usage{
countsBetweenAnchors(x, y, ...)

\S4method{countsBetweenAnchors}{GenomicInteractions,GRanges}(x, y,
  ignore_overlaps = FALSE, ...)
}
\arguments{
\item{x}{A \linkS4class{GenomicInteractions} object.}

\item{y}{A \linkS4class{GenomicRanges} object.}

\item{...}{Extra parameters to pass to \code{\link{findOverlaps}} for \linkS4class{GenomicInteractions} objects.}

\item{ignore_overlaps}{Logical scalar indicating whether overlapping anchors should be disallowed.}
}
\value{
A \linkS4class{GenomicInteractions} object with annotated counts between anchors.
}
\description{
Calculate the number of of paired-end reads mapping between a defined set of anchors.
This function will ignore counts present in the input data.
}
\details{
By default, \code{ignore_overlaps=FALSE} will raise an error when overlapping anchors are observed.
This can be turned off but users should be careful with multi-mapping. 
Setting \code{type="within"} in \code{\link{findOverlaps}} can reduce multi-mapping effects.
}
