% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export.R
\docType{methods}
\name{export.bedpe}
\alias{export.bedpe}
\alias{export.bedpe,GenomicInteractions-method}
\title{Export interactions in BEDPE format.}
\usage{
export.bedpe(GIObject, fn = NULL, score = "counts")

\S4method{export.bedpe}{GenomicInteractions}(GIObject, fn = NULL,
  score = "counts")
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object.}

\item{fn}{String containing a path to a file to write the interaction data to.}

\item{score}{String or integer scalar specifying the metadata column to export as the \code{score} field.}
}
\value{
If \code{fn} is supplied, \code{NULL} is invisibly returned and a file is created at \code{fn}.

Otherwise, a data.frame is returned containing all of the corresponding information from \code{GIObject}.
}
\description{
Exports a \linkS4class{GenomicInteractions} object to file in the BEDPE (paired-end) format,.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)
export.bedpe(hic_example_data, fn = tempfile(), score = "counts")
}
\author{
Malcolm Perry, Elizabeth Ing-Simmons
}
