% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export.R
\docType{methods}
\name{export.chiasig}
\alias{export.chiasig}
\alias{export.chiasig,GenomicInteractions-method}
\title{Export interactions in a BEDPE-like format for use with ChiaSig}
\usage{
export.chiasig(GIObject, fn = NULL, score = "counts")

\S4method{export.chiasig}{GenomicInteractions}(GIObject, fn = NULL,
  score = "counts")
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object.}

\item{fn}{String containing a path to a file to write the interaction data to.}

\item{score}{String or integer scalar specifying the metadata column to export as the \code{score} field.}
}
\value{
If \code{fn} is supplied, \code{NULL} is invisibly returned and a file is created at \code{fn}.

Otherwise, a data.frame is returned containing all of the corresponding information from \code{GIObject}.
}
\description{
Exports a \linkS4class{GenomicInteractions} object to file in a BEDPE-like format for use with ChiaSig.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)
export.chiasig(hic_example_data, fn = tempfile(), score = "counts")

}
\references{
Paulsen et al. (2014).
A statistical model of ChIA-PET data for accurate detection of chromatin 3D interactions.
\emph{Nucleic Acids Res.} 42(18), e143.
}
\author{
Malcolm Perry, Elizabeth Ing-Simmons
}
