% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fugi-package.R
\docType{package}
\name{fugi-package}
\alias{fugi-package}
\title{fugi: Further utitilies for Genomic Interactions}
\description{
This package contains methods that previously were part of the \pkg{GenomicInteractions} package.
They have since been split up, to distinguish the base \linkS4class{GenomicInteractions} class
from the relatively high-level functionality available in this package.
}
\examples{
library(fugi)
}
