% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getters.R
\docType{methods}
\name{name}
\alias{name}
\alias{anchorOne}
\alias{anchorTwo}
\alias{interactionCounts}
\alias{annotationFeatures}
\alias{getters}
\alias{name,GenomicInteractions-method}
\alias{description,GenomicInteractions-method}
\alias{anchorOne,GenomicInteractions-method}
\alias{anchorTwo,GenomicInteractions-method}
\alias{interactionCounts,GenomicInteractions-method}
\alias{annotationFeatures,GenomicInteractions-method}
\alias{seqinfo,GenomicInteractions-method}
\title{Functions to access data in a \linkS4class{GenomicInteractions} object}
\usage{
name(GIObject)

anchorOne(GIObject)

anchorTwo(GIObject)

interactionCounts(GIObject)

annotationFeatures(GIObject)

\S4method{name}{GenomicInteractions}(GIObject)

\S4method{description}{GenomicInteractions}(object)

\S4method{anchorOne}{GenomicInteractions}(GIObject)

\S4method{anchorTwo}{GenomicInteractions}(GIObject)

\S4method{interactionCounts}{GenomicInteractions}(GIObject)

\S4method{annotationFeatures}{GenomicInteractions}(GIObject)

\S4method{seqinfo}{GenomicInteractions}(x)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object.}

\item{object}{Object, possibly derived from class \code{\link{eSet-class}}.}

\item{x}{A \linkS4class{GenomicInteractions} object}
}
\value{
For \code{anchorOne} and \code{anchorTwo}, a \linkS4class{GRanges} for the first and second anchor regions, respectively.

For \code{interactionCounts}, a numeric vector with counts for each interaction in the object.

For \code{description} and \code{name}, a string containing the description and name of the object, respectively.

For \code{annotationFeatures}, a character vector of features with which the object was previously annotated, 
or \code{NA} if the object is unannotated.

For \code{seqinfo}, a \linkS4class{Seqinfo} object for the regions in \code{GIObject}.
}
\description{
These functions can be used to access data stored in the slots of a \linkS4class{GenomicInteractions} object.
They are mostly retained for convenience and backwards compatibility -
users are recommended to use the getters from the \pkg{GenomicInteractions} package itself.
}
\examples{
anchor.one = GRanges(c("chr1", "chr1", "chr1", "chr1"), 
     IRanges(c(10, 20, 30, 20), width=5))
anchor.two = GRanges(c("chr1", "chr1", "chr1", "chr2"), 
     IRanges(c(100, 200, 300, 50), width=5))
interaction_counts = sample(1:10, 4)
test <- GenomicInteractions(anchor.one, anchor.two, 
     counts=interaction_counts)

name(test)
description(test)
anchorOne(test)
anchorTwo(test)
interactionCounts(test)

}
\author{
Malcolm Perry, Elizabeth Ing-Simmons
}
