% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentData.R
\docType{data}
\name{hic_example_data}
\alias{hic_example_data}
\title{Example HiC dataset}
\format{A \linkS4class{GenomicInteractions} object with length 8171.}
\usage{
data(hic_example_data)
}
\value{
A \linkS4class{GenomicInteractions} object.
}
\description{
This dataset contains HiC data from Seitan et al. (2013). 
The data was analysed using HOMER (Heinz et al. 2010) at a resolution of 100kb to find significant interactions. 
This example dataset has been filtered to retain only interactions on chromosomes 14 and 15 with a FDR < 0.1.
The data has also been annotated for overlaps with Refseq promoters.
See the HiC analysis vignette (\code{vignette(package="fugi")}) for more information on how this dataset was created.
}
\references{
Seitan VC et al. (2013).
Cohesin-based chromatin interactions enable regulated gene expression within pre-existing architectural compartments. 
\emph{Genome Res.} 23, 2066-77.

Heinz S, Benner C, Spann N, Bertolino E et al. (2010).
Simple combinations of lineage-determining transcription factors prime cis-regulatory elements required for macrophage and B cell identities. 
\emph{Mol. Cell} 38(4), 576-589.
}
\keyword{datasets}
