% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helpers.R
\docType{methods}
\name{is.pp}
\alias{is.pp}
\alias{is.pd}
\alias{is.pt}
\alias{is.dd}
\alias{is.dt}
\alias{is.tt}
\alias{isInteractionType}
\alias{is.trans}
\alias{is.cis}
\alias{InteractionHelpers}
\alias{is.pp,GenomicInteractions-method}
\alias{is.pd,GenomicInteractions-method}
\alias{is.pt,GenomicInteractions-method}
\alias{is.dd,GenomicInteractions-method}
\alias{is.dt,GenomicInteractions-method}
\alias{is.tt,GenomicInteractions-method}
\alias{isInteractionType,GenomicInteractions-method}
\alias{is.trans,GenomicInteractions-method}
\alias{is.cis,GenomicInteractions-method}
\title{Interaction type helper functions}
\usage{
is.pp(GIObject)

is.pd(GIObject)

is.pt(GIObject)

is.dd(GIObject)

is.dt(GIObject)

is.tt(GIObject)

isInteractionType(GIObject, x, y)

is.trans(GIObject)

is.cis(GIObject)

\S4method{is.pp}{GenomicInteractions}(GIObject)

\S4method{is.pd}{GenomicInteractions}(GIObject)

\S4method{is.pt}{GenomicInteractions}(GIObject)

\S4method{is.dd}{GenomicInteractions}(GIObject)

\S4method{is.dt}{GenomicInteractions}(GIObject)

\S4method{is.tt}{GenomicInteractions}(GIObject)

\S4method{isInteractionType}{GenomicInteractions}(GIObject, x, y)

\S4method{is.trans}{GenomicInteractions}(GIObject)

\S4method{is.cis}{GenomicInteractions}(GIObject)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object.}

\item{x, y}{Names of annotated node classes}
}
\value{
A logical vector indicating whether each entry of \code{GIObject} is of the specified type.
}
\description{
Functions to classify interactions within \linkS4class{GenomicInteractions} objects.
}
\details{
\code{isInteractionType} identifies all interactions that occur between the annotated node classes \code{x} and \code{y}.

\code{is.trans} and \code{is.cis} select trans-chromosomal and intra-chromosomal interactions, respectively.

The other functions are convenience wrappers for identifying interactions between regions of common annotations,
namely promoter \code{p}, distal \code{d} or terminator \code{t} regions.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)
table(is.cis(hic_example_data))
sum(interactionCounts(hic_example_data))

}
\author{
Malcolm Perry, Elizabeth Ing-Simmons
}
