% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionTrack-class.R
\docType{class}
\name{InteractionTrack-class}
\alias{InteractionTrack-class}
\title{The InteractionTrack class}
\description{
A class to hold chromatin interaction data for a specific genomic region.

The InteractionTrack class is a specific \pkg{Gviz}-derived class for enabling the visualisation of chromatin interaction data. 
It allows interactions on a specified chromosome to be visualised by examining interactions between anchors as bezier curves. 
The object is instantiated and used in a similar fashion to standard \pkg{Gviz} tracks, 
e.g., plotted using the \code{\link{plotTracks}} function.

Several additional display parameters are defined for this class:
\describe{
\item{\code{plot.anchors}:}{Logical scalar indicating whether anchor elements are to be drawn.}
\item{\code{col.anchors.line}:}{String specifying the colour of border of these anchor elements.}
\item{\code{col.anchors.fill}:}{String specifying the colur to fill these anchor elements.}
\item{\code{plot.outside}:}{Logical scalar indicating whether interactions spanning outside of the window are to be plotted.}
\item{\code{col.outside}:}{String specifying the line colour to use for outside-spanning interactions.}
\item{\code{plot.trans}:}{Logical scalar indicating whether trans-chromosomal interactions are to be plotted.}
\item{\code{col.trans}:}{String specifying the line colour to use for trans-chromosomal interactions.}
\item{\code{col.interactions}:}{String specifying the colour of arcs representing interactions within the region of interest.}
\item{\code{col.interaction.types}:}{A named character vector of colours to use for each type of interaction,
where the name corresponds to the type of interaction (e.g., promoter-promoter, distal-promoter).}
\item{col.anchors.line.node.class}{A named character vector of colours to use for the border of each type of anchor element,
where the name corresponds to the anchor type (e.g., promoter).}
\item{col.anchors.fill.node.class}{A named character vector of colours to use for the fillof each type of anchor element,
where the name corresponds to the anchor type (e.g., promoter).}
}

A number of options are available to convey quantitative information:
\itemize{
\item By default, the height of an arc representing an interaction is proportional to the number of supporting reads.
Instead of using the counts to define this, the height can be set to be proportional to either \code{fdr} or \code{p.value} using the \code{interaction.measure} display parameter. 
\item By changing the \code{interaction.dimension} to \code{"width"}, the line widths of each arc now represent the statistic supporting them, instead of the heights.
\item The heights of the arcs can be made to be proportional to the log10 of the supporting statistic by changing \code{interaction.dimension.transform} to \code{"log"}. 
}

All of these parameters can be set with standard methods, i.e., \code{\link{displayPars<-}} on a InteractionTrack object.
}
