% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionTrack-class.R
\name{InteractionTrack}
\alias{InteractionTrack}
\title{Constructor to create an InteractionTrack object}
\usage{
InteractionTrack(x, chromosome = "", name = name(x), start = NULL,
  end = NULL)
}
\arguments{
\item{x}{A \linkS4class{GenomicInteractions} object.}

\item{chromosome}{String specifying which chromosome to use to create a track.}

\item{name}{String specifying the name of the track.}

\item{start}{Integer scalar specifying the start location for the track.}

\item{end}{Integer scalar specifying the end location for the track.}
}
\value{
An \linkS4class{InteractionTrack} object.
}
\description{
Create an \linkS4class{InteractionTrack} object from an \linkS4class{GenomicInteractions} object to visualise a specified chromosome.
}
\examples{
library(Gviz)
anchor.one = GRanges(c("chr1", "chr1", "chr1", "chr1"), 
    IRanges(c(10, 20, 30, 20), width=5))
anchor.two = GRanges(c("chr1", "chr1", "chr1", "chr2"), 
    IRanges(c(100, 200, 300, 50), width=5))

interaction_counts = sample(1:10, 4)
test <- GenomicInteractions(anchor.one, anchor.two, counts=interaction_counts)
metadata(test) <- list(description="this is a test", experiment_name="test")
interactions.track = InteractionTrack(name="Test", test, chromosome="chr1")                        
plotTracks(list(interactions.track), chromosome="chr1", from=0, to=500)

}
