% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewPoints.R
\name{plotAvgViewpoint}
\alias{plotAvgViewpoint}
\title{Plot coverage around a set of virtual 4C viewpoints}
\usage{
plotAvgViewpoint(x, left_dist = 1e+05, right_dist = 1e+05,
  ylab = "Average signal", xlab = "Relative position",
  fix = "center", ...)
}
\arguments{
\item{x}{A \linkS4class{GenomicInteractions} object generated by \code{\link{viewPoint}}.}

\item{left_dist}{Integer scalar specifying the distance to the left of each bait to consider, in bp.}

\item{right_dist}{Integer scalar specifying the distance to the right of each bait to consider, in bp.}

\item{ylab}{String containing the Y axis label.}

\item{xlab}{String containing the X axis label.}

\item{fix}{One of \code{"center"}, \code{"start"}, \code{"end"}, to be passed to \code{\link{resize}}.
Interaction distances are calculated relative to this part of the bait.}

\item{...}{Additional arguments to \code{\link{plot}}.}
}
\value{
A coverage track is plotted on the current graphics device.
A \linkS4class{RleList} object containing the aggregated \code{\link{coverage}} is invisibly returned.
}
\description{
Plots summarised coverage of interactions around a set of viewpoints, e.g., promoters. 
This function requires the output of \code{\link{viewPoint}} as input.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)

pos <- GRanges(seqnames="chr15", ranges=IRanges(start=59477709, end=59482708))
region <- GRanges(seqnames="chr15", ranges=IRanges(start=58980209, end=59980208))
vp <- viewPoint(hic_example_data, pos, region)
plotAvgViewpoint(vp, left_dist = 1000000, right_dist = 100000)
}
