% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummaryStats.R
\name{plotCounts}
\alias{plotCounts}
\title{Plot interaction count frequency}
\usage{
plotCounts(GIObject, normalise = FALSE, cut = 10)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object.}

\item{normalise}{Logical scalar specifying whether to plot the proportion of total reads instead of the count.}

\item{cut}{Numeric scalar specifying the upper limit on the count,
above which all interactions are aggregated into a single category.
Setting this to \code{NULL} will not perform any aggregation.}
}
\value{
A ggplot2 plot.
}
\description{
Plot a bar chart of the number of interactions supported by different numbers of reads in a \linkS4class{GenomicInteractions} object.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)

plotCounts(hic_example_data)
plotCounts(hic_example_data, normalise=TRUE)
}
