% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummaryStats.R
\name{plotInteractionAnnotations}
\alias{plotInteractionAnnotations}
\title{Plot interaction types}
\usage{
plotInteractionAnnotations(GIObject, node.classes = NULL,
  viewpoints = NULL, other = 0, keep.order = FALSE, legend = FALSE)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object}

\item{node.classes}{Character vector specifying the node classes to include in the plot.
Defaults to all available node classes.}

\item{viewpoints}{Character vector of acceptable node classes,
where interactions are only considered if least one anchor is of an acceptable node class. 
Defaults to all classes in \code{node.classes}.}

\item{other}{Numeric scalar of the minimum proportion for an interaction type to be shown.
Interaction types making up fewer than \code{other} percent of the total interactions will be consolidated into a single \dQuote{other} category.}

\item{keep.order}{Logical scalar specifying whether the original order of \code{node.classes} should be retained for plotting.
Default is to use the alphabetical order.}

\item{legend}{Logical scalar specifying whether the legend should be plotted.
If \code{FALSE}, donut plot is annotated with category names.}
}
\value{
A ggplot2 plot.
}
\description{
Plot a donut plot of interaction types for an annotated \linkS4class{GenomicInteractions} object.
}
\examples{
library("GenomicRanges")
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)

data(mm9_refseq_promoters)
mm9_refseq_grl = split(mm9_refseq_promoters, mm9_refseq_promoters$id)
annotateInteractions(hic_example_data, list(promoter=mm9_refseq_grl))
plotInteractionAnnotations(hic_example_data)
}
