% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummaryStats.R
\name{plotSummaryStats}
\alias{plotSummaryStats}
\title{Plot summary statistics}
\usage{
plotSummaryStats(GIObject, other = 5, cut = 10)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object.}

\item{other}{Numeric scalar passed to \code{\link{plotInteractionAnnotations}}.}

\item{cut}{Numeric scalar passed to \code{\link{plotCounts}}.}
}
\value{
A plot is made on the current graphics device.
The output of \code{\link{marrangeGrob}} is returned.
}
\description{
Makes summary plots of the counts, interaction distances, interaction annotations, 
and percentage of cis and trans interactions for a \linkS4class{GenomicInteractions} object.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)
plotSummaryStats(hic_example_data)
}
