% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewPoints.R
\name{plotViewpoint}
\alias{plotViewpoint}
\title{Plot coverage around a virtual 4C viewpoint}
\usage{
plotViewpoint(x, region, ylab = "Signal", xlab = NULL, ...)
}
\arguments{
\item{x}{A \linkS4class{GenomicInteractions} object generated by \code{\link{viewPoint}}.}

\item{region}{A \linkS4class{GRanges} object specifying the genomic region to plot.}

\item{ylab}{String containing the Y axis label.}

\item{xlab}{String containing the X axis label. 
By default this is the chromosome of the region that is being plotted.}

\item{...}{Additional arguments to pass to \code{\link{plot}}.}
}
\value{
A coverage track is plotted on the current graphics device.
A \linkS4class{RleList} object containing the result of \code{\link{coverage}} is invisibly returned.
}
\description{
Plots coverage of interactions around a given viewpoint. 
This function requires the output of `viewPoint()` as input. 
You should additionally specify the total region you wish to plot.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)

pos <- GRanges(seqnames="chr15", ranges=IRanges(start=59477709, end=59482708))
region <- GRanges(seqnames="chr15", ranges=IRanges(start=58980209, end=59980208))
vp <- viewPoint(hic_example_data, pos, region)
plotViewpoint(vp, region)

}
