% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{removeDups}
\alias{removeDups}
\title{Remove duplicated interactions}
\usage{
removeDups(GIObject)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object.}
}
\value{
A \linkS4class{GenomicInteractions} object that is a subset of the input object.
}
\description{
Removes all but the first occurence of a duplicated interaction (defined as having identical coordinates for both anchors).
}
\details{
Note that this function will not sum total counts of all the duplicates. 
It is designed for removing potential PCR duplicates after reading in BAM files.
}
