% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_annotation.R
\docType{methods}
\name{resetAnnotations}
\alias{resetAnnotations}
\alias{resetAnnotations,GenomicInteractions-method}
\title{Reset annotations made to a GenomicInteractions object}
\usage{
resetAnnotations(GIObject)

\S4method{resetAnnotations}{GenomicInteractions}(GIObject)
}
\arguments{
\item{GIObject}{An annotated GenomicInteractions object.}
}
\value{
invisible(1)
}
\description{
This function removes all annotations from a \linkS4class{GenomicInteractions} object by
deleting all of the metadata columns associated with both anchors.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)
mcols(regions(hic_example_data, type=1))

resetAnnotations(hic_example_data)
mcols(regions(hic_example_data, type=1))

}
