% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{sameStrand}
\alias{sameStrand}
\title{Check if anchors have the same strand}
\usage{
sameStrand(GIObject)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object.}
}
\value{
A logical vector indicating if both anchors of an interaction are on the same strand.
}
\description{
Tests whether anchors have the same strand, for use in processing paired reads from BAM files.
}
