% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/subsetting.R
\docType{methods}
\name{subsetByFeatures}
\alias{subsetByFeatures}
\alias{subsetByFeatures,GenomicInteractions,GRanges,missing-method}
\alias{subsetByFeatures,GenomicInteractions,GRangesList,missing-method}
\alias{subsetByFeatures,GenomicInteractions,character,character-method}
\title{Subset a \linkS4class{GenomicInteractions} object by features}
\usage{
subsetByFeatures(GIObject, features, feature.class = NULL)

\S4method{subsetByFeatures}{GenomicInteractions,GRanges,missing}(GIObject,
  features, feature.class = NULL)


  \S4method{subsetByFeatures}{GenomicInteractions,GRangesList,missing}(GIObject,
  features, feature.class = NULL)


  \S4method{subsetByFeatures}{GenomicInteractions,character,character}(GIObject,
  features, feature.class = NULL)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object.}

\item{features}{A \linkS4class{GRanges} or \linkS4class{GRangesList} object, 
or a character vector containing IDs of annotated features, e.g. promoter IDs.}

\item{feature.class}{String containing the feature name to use if \code{features} is a character vector.}
}
\value{
A subsetted \linkS4class{GenomicInteractions} object.
}
\description{
Subsets interactions for which at least one of the anchors overlaps with a given \linkS4class{GRanges} object.
Alternatively, subsets interactions based on annotated feature IDs for a particular feature.
}
\examples{
data("hic_example_data")
hic_example_data <- updateObject(hic_example_data)

data("mm9_refseq_promoters")
annotateInteractions(hic_example_data, list(promoter = mm9_refseq_promoters))
ids <- names(mm9_refseq_promoters[1:10])
subsetByFeatures(hic_example_data, ids, "promoter")

}
