% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summarise_annotations.R
\docType{methods}
\name{summariseByFeaturePairs}
\alias{summariseByFeaturePairs}
\alias{summariseByFeaturePairs,GenomicInteractions-method}
\title{Summarise by feature pairs}
\usage{
summariseByFeaturePairs(GIObject, features.one, feature.name.one,
  features.two, feature.name.two)

\S4method{summariseByFeaturePairs}{GenomicInteractions}(GIObject,
  features.one, feature.name.one, features.two, feature.name.two)
}
\arguments{
\item{GIObject}{An annotated \linkS4class{GenomicInteractions} object.}

\item{features.one}{A \linkS4class{GRanges} object containing the feature set of interest.}

\item{feature.name.one}{String containing the name of the first feature set of interest.}

\item{features.two}{A \linkS4class{GRanges} object containing the second feature set of interest.}

\item{feature.name.two}{String containing the name of the second feature set of interest.}
}
\value{
A data.frame with one line for each combination of \code{features.one} and \code{features.two}.
}
\description{
Calculat the number of interactions between two user-provided sets of features. 
This allows the summarisation of the number of features of a specific type that a particular region is involved in,
and how many interactions exist between those features.
}
\examples{
data("hic_example_data")
hic_example_data <- updateObject(hic_example_data)

data("mm9_refseq_promoters")
data("thymus_enhancers")
annotateInteractions(hic_example_data, list(promoter = mm9_refseq_promoters, enhancer = thymus_enh))

# can be slow so subset of features used for examples
regs <- regions(hic_example_data, type=1)
p <- unique(unlist(head(regs$promoter.id)))
e <- unique(unlist(head(regs$enhancer.id)))
p <- p[!is.na(p)]
p <- mm9_refseq_promoters[p]
e <- e[!is.na(e)]
e <- thymus_enh[e]
ep_summary <- summariseByFeaturePairs(hic_example_data, p, "promoter", e, "enhancer")
}
