% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summarise_annotations.R
\docType{methods}
\name{summariseByFeatures}
\alias{summariseByFeatures}
\alias{summariseByFeatures,GenomicInteractions-method}
\title{Summarize by features}
\usage{
summariseByFeatures(GIObject, features, feature.name,
  distance.method = "midpoint", annotate.self = FALSE)

\S4method{summariseByFeatures}{GenomicInteractions}(GIObject, features,
  feature.name, distance.method = "midpoint", annotate.self = FALSE)
}
\arguments{
\item{GIObject}{An annotated \linkS4class{GenomicInteractions} object.}

\item{features}{A \linkS4class{GRanges} object containing the feature set.}

\item{feature.name}{String specifying the name of the feature set.}

\item{distance.method}{String specifying the method for calculating distances between anchors, 
see \code{?\link{calculateDistances}}.}

\item{annotate.self}{Logical scalar indicating whether to annotate self interactions,
i.e., where a feature in `features` overlaps both anchors of an interaction.}
}
\value{
A data.frame with one line for each range in \code{features}.
}
\description{
Compute summary statistics of interactions for a given feature set.
This includes the number of interactions (the sum of all interaction counts), 
number of unique interactions and number of trans (i.e., interchromosomal) interations.  
It also returns some statistics for the distances of interactions for all interactions of the feature, 
and for the different interaction types, e.g., promoter-distal.
}
\examples{
data("hic_example_data")
hic_example_data <- updateObject(hic_example_data)

data("mm9_refseq_promoters")
annotateInteractions(hic_example_data, list(promoter = mm9_refseq_promoters))
summariseByFeatures(hic_example_data, mm9_refseq_promoters[1:10], "promoter")
}
