% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set_annotation.R
\docType{methods}
\name{annotateRegions}
\alias{annotateRegions}
\alias{annotateRegions,GenomicInteractions,character,vector-method}
\title{Annotate regions}
\usage{
annotateRegions(GIObject, name, dat)

\S4method{annotateRegions}{GenomicInteractions,character,vector}(GIObject,
  name, dat)
}
\arguments{
\item{GIObject}{A GenomicInteractions or GenomicInteractions object}

\item{name}{Character. Will be used as a column name.}

\item{dat}{Vector of the same length as the GenomicInteractions object,
containing data with which to annotate the object.}
}
\value{
invisible(1)
}
\description{
Use this function to add metadata parallel to the `regions` slot of a 
GenomicInteractions or GenomicInteractions object.
}
\examples{
data(hic_example_data)
hic_example_data <- updateObject(hic_example_data)
mcols(regions(hic_example_data, type=1))

chip <- runif(length(regions(hic_example_data, type=1)), max=1000)
annotateRegions(hic_example_data, "chip", chip)
mcols(regions(hic_example_data, type=1))

}
