% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calculateDistances.R
\docType{methods}
\name{calculateDistances}
\alias{calculateDistances}
\alias{calculateDistances,GenomicInteractions-method}
\title{Calculate interaction distances}
\usage{
calculateDistances(GIObject, method = "midpoint", floor = TRUE)

\S4method{calculateDistances}{GenomicInteractions}(GIObject,
  method = "midpoint", floor = TRUE)
}
\arguments{
\item{GIObject}{A \linkS4class{GenomicInteractions} object}

\item{method}{Character vector indicating how to calculate distances, must
be one of `midpoint', `outer', `inner'.}

\item{floor}{A logical specifying whether to round down distances to nearest 
base pair or not. Default TRUE.}
}
\value{
An vector containing the distances between anchors/GRanges,
        NA if on different chromosomes, rounded down to the nearest bp.
}
\description{
This will be deprecated in favor of the \code{\link{pairdist}} function from \pkg{GenomicInteractions}.
}
\examples{
anchor.one <- GRanges(c("chr1", "chr1", "chr1", "chr1"), 
    IRanges(c(10, 20, 30, 20), width=5))
anchor.two <- GRanges(c("chr1", "chr1", "chr1", "chr2"), 
    IRanges(c(100, 200, 300, 50), width=5))
test <- GenomicInteractions(anchor.one, anchor.two)
calculateDistances(test, method="midpoint")
        
}
